#Conflicts:
#	Carry1st/Repository/CartRepository.swift

+++ Carry1st/Features/Cart/cartCheckOut/CheckOutScreen.swift
@@ -12,12 +12,13 @@ struct CheckOutScreen: View {
     @State private var itemCount: Int = 1
     @State private var couponCode: String = ""
     @Environment(\.presentationMode) var presentationMode
+    var isBuyNow: Bool? = false
+    var buyNowData: ProductData?
     @Query var tasks: [ProductData]
     @Environment(\.modelContext) var context
-    
+    @StateObject private var delegateHandler = CartDelegateHandler()
     var body: some View {
         VStack{
-            
             NormalBackAndTitleAppBar(title: "Checkout", backAction: {
                 presentationMode.wrappedValue.dismiss()
             })
@@ -26,23 +27,32 @@ struct CheckOutScreen: View {
               name: "Michael Ossai", address: "1 Sunday Ogunyade Street, Gbagada Express Way, beside Eterna Fuel Station, Gbagada, Lagos 100234, Nigeria"
             )
             HDivider()
-            
-            List(tasks) { item in
-                       CartItemWidget(
-                           count: item.quantity,
-                           productItem: item, onQuantityChange: {
-                               count in
-                               if count < 1 {
-                                   context.delete(item)
-                                   return
-                               }
-                               CartUtils.updateCounts(context: context, data: item, quantity: count)
-                           }
-                       )
-                       .buttonStyle(PlainButtonStyle())
-                   }
-                   .listStyle(PlainListStyle())
-            
+            if isBuyNow == true {
+                CartItemWidget(
+                    count: buyNowData!.quantity,
+                    productItem: buyNowData!, onQuantityChange: {
+                        count in
+                       
+                    }
+                )
+                Spacer()
+            }else {
+                List(tasks) { item in
+                    CartItemWidget(
+                        count: item.quantity,
+                        productItem: item, onQuantityChange: {
+                            count in
+                            if count < 1 {
+                                context.delete(item)
+                                return
+                            }
+                            CartService.shared.updateItemQuantity(context: context, data: item, quantity: count)
+                        }
+                    )
+                    .buttonStyle(PlainButtonStyle())
+                }
+                .listStyle(PlainListStyle())
+            }
             VStack {
                 HStack {
                     HStack {
@@ -64,7 +74,7 @@ struct CheckOutScreen: View {
                 
                 
                 TotalAndDeliveryText(
-                    subTotal: CartUtils.calculateSubTotal(tasks: tasks), deliveryFee: "$5.00", total: CartUtils.calculateTotal(tasks: tasks)
+                    subTotal:  CartService.shared.calculateSubTotal(tasks: tasks), deliveryFee: "$5.00", total:  CartService.shared.calculateTotal(tasks: tasks)
                 )
                 NavigationButton(
                     destination: SuccessFulOrderScreen(), action: {
@@ -74,10 +84,23 @@ struct CheckOutScreen: View {
                 }
             }
         }.background(Color.white)
+            .onAppear {
+                CartService.shared.delegate = delegateHandler
+            }
             .navigationBarHidden(true)
+            .toastView(toast: $delegateHandler.toast)
     }
 }
 
+extension CheckOutScreen: CartServiceProtocol {
+    func cartActionSuccessful(msg: String) {
+        
+    }
+    
+    func cartActionFailed(msg: String) {
+        
+    }
+}
 #Preview {
     CheckOutScreen()
 }
